LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-marker-controls-scroll-tracking-001-expected.txt


FAIL ::scroll-marker is not activated when its originating element is not scrolled into the view assert_equals: ::scroll-marker background-color should be rgb(255, 0, 0) but was  expected "rgb(255, 0, 0)" but got ""
FAIL ::scroll-marker is activated when its originating element is scrolled into the view assert_equals: ::scroll-marker background-color should be rgb(0, 128, 0) but was  expected "rgb(0, 128, 0)" but got ""
